requirements.txt
Flask
Pillow
flask-cors
tensorflow
numpy
